Move all druid sections to components folder and remove links to GitHub in footer.
