+++ src/components/Druid/Sections/BoonsAndBuffs/Spirits.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import WikiIcon from '../../../components/WikiIcon'
+import WikiIcon from '../../../WikiIcon'
 
 const Spirits = () => {
     return (
