+++ src/components/Druid/Sections/index.js
@@ -1,0 +1,6 @@
+import BoonsAndBuffs from './BoonsAndBuffs'
+import Rotations from './Rotations'
+import Traits from './Traits'
+import SkillsAndUtilities from './SkillsAndUtilities'
+
+export {BoonsAndBuffs, Rotations, Traits, SkillsAndUtilities}
