--- src/pages/Druid/BoonsAndBuffs/index.js
@@ -1,5 +1,0 @@
-import Boons from './Boons';
-import Spirits from './Spirits';
-import Spotter from './Spotter';
-
-export {Boons, Spirits, Spotter};
