--- src/pages/Druid/Rotations/index.js
@@ -1,5 +1,0 @@
-import Openers from './Openers'
-import WeaponSwaps from './WeaponSwaps'
-import BossRotations from './BossRotations'
-
-export {Openers, WeaponSwaps, BossRotations};
