--- src/pages/Druid/Traits/index.js
@@ -1,6 +1,0 @@
-import Skirmishing from './Skirmishing';
-import NatureMagic from './NatureMagic';
-import Druid from './Druid';
-import Marksmanship from './Marksmanship';
-
-export {Skirmishing, NatureMagic, Druid, Marksmanship};
